new feature: NBT decompilation to XML
Initial implementation of decompilation functionality from NBT data to
XML format. This uses `quick_xml` crate to write formatted XML data.

This change also removes modules `util` and `state`. The api is mostly
unused in the mentioned modules, with the some of the functions in
`util` being moved to new module `cmd`. The new module will contain the
main business logic for processing commands, by having function for each
command that the cli supports. As for `state` - it was best to remove
for now as it really is not needed for basic commands such as `Compile`
and `Decompile` as they don't require complex state management and may
overcomplicate the IO state manager api if its usage is required in
those commands.

Closes #6