cli: rework input/output resolving
Changes the output to default directly into STDOUT. Default behaviour
can be overwritten if user specifies `--file` or `-f` short-hand. This
is meant to simplify the logic for outputing the result which then can
be processed the way user wishes to. A great example would be with
`tar` command man page.

This also implements ability to specify the source of input file. Input
source can be: FILE (/path/to/file), "-" (STDIN), empty (STDIN). Another
example would `cat` command man page.

Closes #14